requirements.txt
pandas
numpy
matplotlib
torch
ta
requests
python-dotenv
python-dateutil
jupyter
streamlit
